Refactor to get rid of app directory
