+++ lib/main_rdt.dart
@@ -2,19 +2,19 @@ import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:crowdleague/middleware/app_middleware.dart';
+import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/reducers/app_reducer.dart';
+import 'package:crowdleague/services/auth_service.dart';
 import 'package:crowdleague/services/conversations_service.dart';
 import 'package:crowdleague/services/leaguers_service.dart';
 import 'package:crowdleague/services/navigation_service.dart';
 import 'package:crowdleague/services/notifications_service.dart';
+import 'package:crowdleague/utils/apple_signin_object.dart';
+import 'package:crowdleague/widgets/crowd_league_app.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:crowdleague/models/app/app_state.dart';
-import 'package:crowdleague/services/auth_service.dart';
-import 'package:crowdleague/utils/apple_signin_object.dart';
-import 'package:crowdleague/app/app.dart';
 import 'package:redux/redux.dart';
 import 'package:redux_remote_devtools/redux_remote_devtools.dart';
