+++ test_driver/app.dart
@@ -1,4 +1,3 @@
-import 'package:crowdleague/app/app.dart';
 import 'package:crowdleague/middleware/app_middleware.dart';
 import 'package:crowdleague/middleware/auth_middleware.dart';
 import 'package:crowdleague/middleware/navigation_middleware.dart';
@@ -6,6 +5,7 @@ import 'package:crowdleague/middleware/notifications_middleware.dart';
 import 'package:crowdleague/models/app/app_state.dart';
 import 'package:crowdleague/reducers/app_reducer.dart';
 import 'package:crowdleague/services/navigation_service.dart';
+import 'package:crowdleague/widgets/crowd_league_app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:redux/redux.dart';
